requirements_from_venv.txt
absl-py
accelerate
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
antropy
anyio
attrs
audioread
blinker
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
coloredlogs
colorlog
contourpy
cramjam
cycler
dash
dash-bootstrap-components
databricks-sdk
datasets
dill
docker
efficientnet-pytorch
einops
fastapi
fastparquet
filelock
flask
flatbuffers
fonttools
frozenlist
fsspec
GitPython
gitdb
google-auth
gputil
graphene
graphql-core
graphql-relay
greenlet
grpcio
gunicorn
h11
h5py
huggingface-hub
humanfriendly
idna
imageio
imbalanced-learn
importlib-metadata
itsdangerous
jinja2
joblib
jsonschema
jsonschema-specifications
kaggle
kiwisolver
kornia
lazy-loader
librosa
lightgbm
lime
llvmlite
mako
markdown
MarkupSafe
matplotlib
mlflow
mne
mpmath
msgpack
multidict
multiprocess
narwhals
networkx
numba
numpy
onnx
onnxruntime
onnxruntime-gpu
opencv-python
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
optuna
packaging
pandas
patsy
pillow
platformdirs
plotly
pooch
propcache
protobuf
psutil
pyarrow
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-core
pyedflib
pyparsing
python-dateutil
pytz
PyWavelets
PyYAML
ray
redis
referencing
regex
requests
retrying
rpds-py
rsa
safetensors
scikit-image
scikit-learn
scipy
seaborn
sentry-sdk
setproctitle
shap
six
sklearn-compat
slicer
smmap
sniffio
soundfile
soxr
sqlalchemy
sqlparse
starlette
statsmodels
sympy
tensorboard
tensorboard-data-server
threadpoolctl
tifffile
timm
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
triton
typing-extensions
typing-inspection
tzdata
urllib3
uvicorn
wandb
werkzeug
xgboost
xxhash
yarl
zipp 